+++ src/internal/dots.h
@@ -17,6 +17,7 @@ enum dots_expansion_op {
   OP_EXPR_UQN,
   OP_EXPR_FIXUP,
   OP_EXPR_DOT_DATA,
+  OP_EXPR_CURLY,
   OP_QUO_NONE,
   OP_QUO_UQ,
   OP_QUO_UQE,
@@ -24,6 +25,7 @@ enum dots_expansion_op {
   OP_QUO_UQN,
   OP_QUO_FIXUP,
   OP_QUO_DOT_DATA,
+  OP_QUO_CURLY,
   OP_VALUE_NONE,
   OP_VALUE_UQ,
   OP_VALUE_UQE,
@@ -31,6 +33,7 @@ enum dots_expansion_op {
   OP_VALUE_UQN,
   OP_VALUE_FIXUP,
   OP_VALUE_DOT_DATA,
+  OP_VALUE_CURLY,
   OP_DOTS_MAX
 };