+++ src/internal/utils.h
@@ -5,6 +5,7 @@
 sexp* new_preserved_empty_list();
 void signal_soft_deprecated(const char* msg);
 sexp* rlang_ns_get(const char* name);
+sexp* rlang_enquo(sexp* sym, sexp* frame);
 
 
 #endif