+++ tests/testthat/test-quasiquotation.R
@@ -581,3 +581,34 @@ test_that("splicing language objects still works", {
 test_that("can unquote string in function position", {
   expect_identical_(expr((!!"foo")()), quote("foo"()))
 })
+
+test_that("{{ is a quote-unquote operator", {
+  fn <- function(foo) expr(list({{ foo }}))
+  expect_identical_(fn(bar), expr(list(!!quo(bar))))
+  expect_identical_(expr(list({{ letters }})), expr(list(!!quo(!!letters))))
+  expect_error_(expr(list({{ quote(foo) }})), "must be a symbol")
+})
+
+test_that("{{ only works in quoting functions", {
+  expect_error_(
+    list2({{ "foo" }}),
+    "Can't use `{{` in a non-quoting function",
+    fixed = TRUE
+  )
+})
+
+test_that("{{ on the LHS of :=", {
+  foo <- "bar"
+  expect_identical_(exprs({{ foo }} := NA), exprs(bar = NA))
+
+  foo <- quote(bar)
+  expect_identical_(exprs({{ foo }} := NA), exprs(bar = NA))
+
+  foo <- quo(bar)
+  expect_identical_(exprs({{ foo }} := NA), exprs(bar = NA))
+
+  fn <- function(foo) exprs({{ foo }} := NA)
+  expect_identical_(fn(bar), exprs(bar = NA))
+
+  expect_error_(exprs({{ foo() }} := NA), "must be a symbol")
+})